Update data.md
Hi  I was using your guide and thanks its a lot of work! Im using windows and making a multiple boot with hackintosh on a different drive with data, for that I follow the instructions but when using CMD not all the time the new EFI partition is selected by default, so maybe adding the command select part x will help folks that are using this method! Regards!